+++ collections/comments.js
@@ -59,6 +59,9 @@ Meteor.methods({
 
     var newCommentId=Comments.insert(comment);
 
+    // increment comment count
+    Meteor.users.update({_id: userId}, {$inc: {commentCount: 1}});
+
     // extend comment with newly created _id
     comment = _.extend(comment, {_id: newCommentId});
 
@@ -123,6 +126,10 @@ Meteor.methods({
     if(canEdit(Meteor.user(), comment)){
       // decrement post comment count
       Posts.update(comment.post, {$inc: {comments: -1}});
+
+      // decrement user comment count
+      Meteor.users.update({_id: comment.userId}, {$inc: {commentCount: -1}});
+
       // note: should we also decrease user's comment karma ?
       Comments.remove(commentId);
     }else{